==> 01_PyCode/22_AN_FirstTestRandomForest_v1.ipynb <==
# First Test of Random Forest

-*- coding: utf-8 -*-

Created on Fri May 20 12:18:05 2022

@author: li.chao.987@s.kyushu-u.ac.jp


import glob
import pandas as pd
import numpy as np
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import RepeatedKFold
from sklearn.ensemble import RandomForestRegressor

otherDataFolder = "F:\\17_Article\\01_Data\\99_MiddleFileStation\\"

df_merged = pd.read_pickle(otherDataFolder+"99_mergedDataset2015.pkl")

df_merged_used = df_merged.iloc[:, 4:54]
df_merged_used.shape

df_merged_used = df_merged_used.dropna()
X = df_merged_used.iloc[:, 1:50]
y = df_merged_used.iloc[:, 0:1]

model = RandomForestRegressor(n_estimators=1000, oob_score=True, random_state=1, n_jobs=-2)

model.fit(X, y)

model.oob_score_
y_pred = model.predict(X)

from sklearn.metrics import r2_score
r2 = r2_score(y, y_pred)
r2

model.oob_score_

from sklearn.model_selection import train_test_split
Xtrain, Xtest, ytrain, ytest = train_test_split(X, y, train_size = 0.8,
                                                random_state=1)


model_cv = RandomForestRegressor(n_estimators=1000, oob_score=True, random_state=1, n_jobs=-2)

model_cv.fit(Xtrain, ytrain)

yhat = model_cv.predict(Xtest)

from sklearn.metrics import r2_score
r2 = r2_score(ytest, yhat)
r2
